move project functionality from project_unauthUsers.jsx to Project.jsx. Delete Project_unauthUsers.jsx file
