+++ src/App.jsx
@@ -9,7 +9,6 @@ import CreateProfile from './components/Profile/CreateProfile.jsx';
 import EditProfile from './components/Profile/EditProfile.jsx';
 import ProjectsList from './components/Project/ProjectsList.jsx';
 import Project from './components/Project/Project.jsx';
-import Projects from './components/Project/Project_unauthUsers.jsx';
 import CreateProject from './components/Project/CreateProject.jsx';
 import UserProjects from './components/Project/UserProjects.jsx';
 import Notification from './components/Notification/Notification.jsx';
@@ -31,8 +30,7 @@ function App() {
           <Route path="/register" element={<Register />} />
           <Route path="/search" element={<Search />} />
           <Route path="/search-results" element={<SearchResults />} />
-          <Route path="/project" element={<Project />} />
-          <Route path="/projects/:projectId" element={<Projects />} />
+          <Route path="/projects/:projectId" element={<Project />} />
           <Route path="/profile" element={<ProtectedRoute element={<Profile />} />} />
           <Route path="/create-profile" element={<ProtectedRoute element={<CreateProfile />} />} />
           <Route path="/edit-profile" element={<ProtectedRoute element={<EditProfile />} />} />
